linux: add linux-milkv-megrez-dev recipe
this kernel uses the rockos upstream, which is what i initially tested
on my megrez. the defconfig is taken from rockos running on my megrez.

i opted to disable the broadcom wifi driver. i have the wifi card, but
didn't want to take time to fix the build for now as it is not critical
to me, and it's an out of tree driver, that will also probably require
firmware to work.

Signed-off-by: Nick Owens <mischief@offblast.org>