tftp-mmc-boot.txt: add Image + fdtcontroladdr boot step
For the Nezha, since we're now using the Image kernel format, rely on
the fdtcontroladdr holding the location of the dtb and add a boot option
in tftp-mmc-boot.txt for that case. This allows the board to boot after
falling through the other options. A longer-term solution should further
simplify this file since we know it won't be using a uImage and may not
need the other configs by default.

Signed-off-by: Trevor Gamblin <tgamblin@baylibre.com>